Camera Auto Mode added
1. several util math functions
2. Automator (automatically interpolate and switch)

+++ Renderer/TerrainRenderer/EngineProperty.hpp
@@ -21,10 +21,10 @@ constexpr float CAMERA_MAX_DEPTH	= 3000.0f;
 constexpr bool	ENABLE_CULLING		= true;
 constexpr float MOUSE_SENSITIVITY	= 0.1f;
 constexpr float CAMERA_SPEED		= 80.0f;
-constexpr float REFLECTION_WIDTH	= CLIENT_WIDTH;
-constexpr float REFLECTION_HEIGHT	= CLIENT_HEIGHT;
-constexpr float REFRACTION_WIDTH	= CLIENT_WIDTH;
-constexpr float REFRACTION_HEIGHT	= CLIENT_HEIGHT;
+constexpr int	REFLECTION_WIDTH	= CLIENT_WIDTH;
+constexpr int	REFLECTION_HEIGHT	= CLIENT_HEIGHT;
+constexpr int	REFRACTION_WIDTH	= CLIENT_WIDTH;
+constexpr int	REFRACTION_HEIGHT	= CLIENT_HEIGHT;
 
 namespace Color {
